Moved base modelers into single module
